feat!: html regex
This refactor will allow us to implement filtering based on element.tagName and perform static analysis on the entire element.

This change did come at a cost though: performance. The regex is not that heavy, but iterating over that data is. We need to lower our memory allocations and improve the escape function, or make it so we don't need to escape at all (rebuild the selector instead of re-using it?)